helm-toolkit: extend ingress annotations with cert-manager
The examples show using issuerRef under the tls key under default or
public in a number of places to override settings related to
cert-manager but those don't uniformly take effect. This makes it work
uniformly on the ingresses as well as letting any values set on the
default take precedence. Lastly this adds cert-manager annotations to
the <svc>-cluster-fqdn and <svc>-namespace-fqdn Ingresses.

Change-Id: Ia114e587b096174fa3154d3bcb85dce6841fe10a
Signed-off-by: Doug Goldstein <doug.goldstein@rackspace.com>